Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-multipart
pyjwt[crypto]
bcrypt
pydantic_settings
pydantic[email]